refactor(core-weave): extract RDF and Turtle helpers
- move core weave RDF parser/query/resolution helpers into rdf_helpers.ts
- move Turtle subject-block editing helpers into turtle_blocks.ts
- keep source-locator helpers, SFLO semantic shorthands, shape assertions, slice classification, and renderer behavior in weave.ts
- update the codebase overview with the extracted helper modules
